feat(web): add timestamps and active state to home feed notification rows
Show relative timestamps (e.g. '2h ago', '1w ago') on notification rows
by default, replacing them with action icons on hover — matching macOS
client behavior. Add visual active state for the currently selected
notification row.

Also moves formatRelativeDate from domains/settings/ to shared
utils/format-date.ts to avoid cross-domain imports.
